Update README.md with project info
Starting to fill out the project with plans.

Going to need to do some research now on Delucia itself and plot out some points to mine.

I think the biggest issue I am going to have right now is getting a server up and running because my RunUO server files are not working. May just spin up a Windows instance just to make it easy on myself instead of trying to get it to work with Mono.