Update README.md with mining info
Talked about the sub-processes for mining and what we are likely and unlikely to automate.

Going to need to now go through the mining locations and talking through the 8x8 grid. There is a good image on uo.com that I saved for Delucia. It appears that, if we use the trinsic passage gate (bottom left) location of 5148/4063 and the magic hut (top right) location of 5295/3979, then we get a workable area of roughly 150x85. That gives us 200 or so 8x8 plots to mine, which should be more than enough hopefully.