Adding research relating to other areas.
Did some more research as I am not sure Delucia is the best location to be doing this sort of thing. I am going to do it though just to stay faithful to the project that I had in mind, but other places, specifically Cove, Wind, and Bucs Den, provide inviting targets as well to check out if I were to extend this further.