Continuing working through 1.0 tasks.
I have banking, restocking, and walking done. Need to get mining,
making tools, and combining ore done before this can be considered a 1.0
candidate.

One thing that I am worried about that I have not yet really worked
through is how to deal with having way too much weight on the character
during the runs. I think I might have to handle pack animals in order to
handle the amount of weight that is going to be mined. Honestly, I am
not even sure if that is going to be enough to deal with it. I will need
to do out the math to see whether or not 5 pack animals (llamas or
horses, don't know if there is a difference) is going to be able to do
it. I could also consider giant beetles, but those were not around
during the time period that I am attempting to be accurate to.