1.0 Done! (Kinda)
So the script works. I am able to get through about 6-7 different spots
without issues, but that is where I run out of packhorse space because
I only have the one coded right now.

Next steps are to add more horse power! (ha)

There is a way that I could be really clever by going through and doing
combinations of ore within packs and also paying attention to different
weight levels that each individual horse is carrying. I am not going to
do any of that noise though because EasyUO is garbage and I want to be
done with this already. Instead, we know that we have exactly 30 spots
that we are mining from. A single horse can carry ore from 6 spots it
seems. So I am going to make it so that there is a switch for which
horse to give the loot to based on which spot we are currently on.