Updated amount of ingots we need to 50.
I still don't actually know if this is enough, but it's definitely in
the right direction. Might need closer to 70 that's a multiple of 4.

Also getting a weird issue where it's trying to use a serial that's N/A.
I think it's because of a weird case when it's attempting to combine ore
from the backpack and it finds an N/A serial pile?