Made changes to oreCombine to try to fix crashing.
Still running into the issue where it's crashing. I believe this is
because it's attempting to find small piles when it should only be
looking for one of the three larger pile types to combine with the
smaller type.

A silly fix to this would be to have a separate list that only includes
the larger ore graphic types as that would remove the possibility
altogether.