Version 1.1.
This should be the one if I am lucky.

There are definitely still things that can be done to make this better,
but this should be good enough such that it can do complete loops by
itself.

These specific changes fix some issues that have been having, such as it
attempting to use indexes of ore when combining that are likely not
existent which was causing client crashes. Another problem is with the
pause relating to the tinkering gump not being long enough when clicking
next page so sometimes it's clicking on the sewing kit.

I also ran into an issue that was not code related. I had it where the
client was on always run. This speeds up the player's movement for sure,
but it also leaves the pack animals in the dust on the way back to the
bank.